Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yfinance>=0.2.50
httpx>=0.27.0
python-dotenv>=1.0.0
cachetools>=5.5.0
google-generativeai>=0.8.0
pydantic>=2.10.0
motor>=3.7.0
pymongo>=4.16.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.22
tenacity>=9.0.0
slowapi>=0.1.9
redis>=5.2.0
prometheus-fastapi-instrumentator>=7.0.0
python-json-logger>=2.0.7
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
ruff>=0.8.0
